small tweaks before 0.3.0

+++ lib/subspace/commands/base.rb
@@ -29,6 +29,12 @@ module Subspace
       end
 
       def template(src, dest = nil, render_binding = nil)
+        return unless confirm_overwrite File.join(dest_dir, dest || src)
+        template! src, dest, render_binding
+        say "Wrote #{dest}"
+      end
+
+      def template!(src, dest = nil, render_binding = nil)
         dest ||= src
         template = ERB.new File.read(File.join(template_dir, "#{src}.erb")), nil, '-'
         File.write File.join(dest_dir, dest), template.result(render_binding || binding)
@@ -36,7 +42,15 @@ module Subspace
 
       def copy(src, dest = nil)
         dest ||= src
+        return unless confirm_overwrite File.join(dest_dir, dest)
         FileUtils.cp File.join(template_dir, src), File.join(dest_dir, dest)
+        say "Wrote #{dest}"
+      end
+
+      def confirm_overwrite(file_path)
+        return true unless File.exists? file_path
+        answer = ask "#{file_path} already exists. Reply 'y' to overwrite: [no] "
+        return answer.downcase.start_with? "y"
       end
 
     end
